+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,94 @@
 
 ### Patch Changes
 
+- Updated dependencies [1c2a4c1]
+  - ai@5.1.0-beta.13
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- ai@5.1.0-beta.12
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [4616b86]
+  - ai@5.1.0-beta.11
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [8c98371]
+  - ai@5.1.0-beta.10
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [ed329cb]
+- Updated dependencies [177b475]
+- Updated dependencies [522f6b8]
+  - ai@5.1.0-beta.9
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [7eca093]
+  - ai@5.1.0-beta.8
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [5a4e732]
+  - ai@5.1.0-beta.7
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [0c4822d]
+  - ai@5.1.0-beta.6
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- ai@5.1.0-beta.5
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- ai@5.1.0-beta.4
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- ai@5.1.0-beta.3
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- ai@5.1.0-beta.2
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [a7f6f81]
+  - ai@5.1.0-beta.1
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
 - Updated dependencies [78928cb]
   - ai@5.1.0-beta.0